Enabled swaping weapons and armors from the inventory. Enabled using a potion. Implemented a chain of responsibility to handle item usage.

+++ Project/Source/Domains/Inventory/Inventory.cpp
@@ -1,8 +1,10 @@
 #include "Inventory.h"
+#include "ItemUsage/ItemHandlerService.h"
 
 Inventory::Inventory()
 {
 	inventoryData = std::make_unique<InventoryDataModel>();
+	itemHandlerService = std::make_unique<ItemHandlerService>();
 }
 
 Inventory::~Inventory()
@@ -18,15 +20,25 @@ void Inventory::StoreItem(std::shared_ptr<Item> item)
 
 void Inventory::RemoveItem(std::shared_ptr<Item> item)
 {
-	inventoryData->DeleteItem(item);
+	if (item->GetStacks() == 1)
+	{
+		inventoryData->DeleteItem(item);
+	}
+	else
+	{
+		item->SetStacks(item->GetStacks() - 1);
+	}
 	SignalInventoryUpdate();
 }
 
 bool Inventory::UseItem(std::shared_ptr<Item> item, EntityStatsDataModel* statsData)
 {
-	//TODO: Write a chain of reponsibility for consuming/equiping items.
-	statsData->SetHealthPoints(statsData->GetHealthPoints() + item->GetModifier());
-	return true;
+	if (itemHandlerService->Handle(item, statsData))
+	{
+
+		return true;
+	}
+	return false;
 }
 
 std::vector<std::shared_ptr<Item>> Inventory::GetItemsByType(ItemType type)
